Separando input component para seu proprio arquivo
