+++ src/components/input/input-props.d.ts
@@ -1,0 +1,8 @@
+export interface InputProps {
+    icon?: any;
+    label?: string;
+    placeholder?: string;
+    value: any;
+    onChangeText: (value: any) => void;
+    secureTextEntry?: boolean;
+}
